+++ packages/plugins/typescript/graphql-request/src/config.ts
@@ -22,4 +22,17 @@ export interface RawGraphQLRequestPluginConfig extends RawClientSideBasePluginCo
    * ```
    */
   rawRequest?: boolean;
+
+  /**
+   * @description Allows you to override the type for extensions when `rawRequest` is enabled.
+   * @default any
+   *
+   * @exampleMarkdown
+   * ```yml
+   * config:
+   *   rawRequest: true
+   *   extensionsType: unknown
+   * ```
+   */
+  extensionsType?: string;
 }