+++ packages/plugins/typescript/react-query/src/fetcher-fetch.ts
@@ -1,4 +1,9 @@
-import { generateInfiniteQueryKey, generateQueryKey, generateQueryVariablesSignature } from './variables-generator';
+import {
+  generateInfiniteQueryKey,
+  generateMutationKey,
+  generateQueryKey,
+  generateQueryVariablesSignature,
+} from './variables-generator';
 
 import { FetcherRenderer } from './fetcher';
 import { OperationDefinitionNode } from 'graphql';
@@ -96,7 +101,8 @@ function fetcher<TData, TVariables>(endpoint: string, requestInit: RequestInit, 
     documentVariableName: string,
     operationName: string,
     operationResultType: string,
-    operationVariablesTypes: string
+    operationVariablesTypes: string,
+    hasRequiredVariables: boolean
   ): string {
     const variables = `variables?: ${operationVariablesTypes}`;
     const hookConfig = this.visitor.queryMethodMap;
@@ -113,6 +119,7 @@ function fetcher<TData, TVariables>(endpoint: string, requestInit: RequestInit, 
       ${options}
     ) =>
     ${hookConfig.mutation.hook}<${operationResultType}, TError, ${operationVariablesTypes}, TContext>(
+      ${generateMutationKey(node)},
       (${variables}) => fetcher<${operationResultType}, ${operationVariablesTypes}>(dataSource.endpoint, dataSource.fetchParams || {}, ${documentVariableName}, variables)(),
       options
     );`;