+++ packages/plugins/typescript/typescript/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @graphql-codegen/typescript
 
+## 2.3.1
+
+### Patch Changes
+
+- 6c898efe5: list all dependencies used by the package in the package.json
+- Updated dependencies [f3833243d]
+- Updated dependencies [6c898efe5]
+  - @graphql-codegen/schema-ast@2.4.0
+
 ## 2.3.0
 
 ### Minor Changes