+++ packages/plugins/typescript/urql-graphcache/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @graphql-codegen/typescript-urql-graphcache
 
+## 2.2.1
+
+### Patch Changes
+
+- 873d409cd: Make typescript-urql-graphcache respect the `useTypeImports` option.
+- f946ffb55: Fixes a bug where the mutation root type name was hardcoded as `Mutation` creating incorrectly named types. Now, the correct mutation type as defined in the schema is used.
+
 ## 2.2.0
 
 ### Minor Changes